+++ packages/client/src/devices/SpeakerManager.ts
@@ -34,23 +34,31 @@ export class SpeakerManager {
    * Lists the available audio output devices
    *
    * Note: It prompts the user for a permission to use devices (if not already granted)
+   * Note: This method is not supported in React Native
    *
    * @returns an Observable that will be updated if a device is connected or disconnected
    */
   listDevices() {
+    if (isReactNative()) {
+      throw new Error(
+        'This feature is not supported in React Native. Please visit https://getstream.io/video/docs/reactnative/core/camera-and-microphone/#speaker-management for more details',
+      );
+    }
     return getAudioOutputDevices();
   }
 
   /**
    * Select a device.
    *
-   * Note: this method is not supported in React Native
+   * Note: This method is not supported in React Native
    *
    * @param deviceId empty string means the system default
    */
   select(deviceId: string) {
     if (isReactNative()) {
-      throw new Error('This feature is not supported in React Native');
+      throw new Error(
+        'This feature is not supported in React Native. Please visit https://getstream.io/video/docs/reactnative/core/camera-and-microphone/#speaker-management for more details',
+      );
     }
     this.state.setDevice(deviceId);
   }
@@ -63,11 +71,13 @@ export class SpeakerManager {
    * Set the volume of the audio elements
    * @param volume a number between 0 and 1.
    *
-   * Note: this method is not supported in React Native
+   * Note: This method is not supported in React Native
    */
   setVolume(volume: number) {
     if (isReactNative()) {
-      throw new Error('This feature is not supported in React Native');
+      throw new Error(
+        'This feature is not supported in React Native. Please visit https://getstream.io/video/docs/reactnative/core/camera-and-microphone/#speaker-management for more details',
+      );
     }
     if (volume && (volume < 0 || volume > 1)) {
       throw new Error('Volume must be between 0 and 1');
@@ -78,14 +88,16 @@ export class SpeakerManager {
   /**
    * Set the volume of a participant.
    *
-   * Note: this method is not supported in React Native.
+   * Note: This method is not supported in React Native.
    *
    * @param sessionId the participant's session id.
    * @param volume a number between 0 and 1. Set it to `undefined` to use the default volume.
    */
   setParticipantVolume(sessionId: string, volume: number | undefined) {
     if (isReactNative()) {
-      throw new Error('This feature is not supported in React Native');
+      throw new Error(
+        'This feature is not supported in React Native. Please visit https://getstream.io/video/docs/reactnative/core/camera-and-microphone/#speaker-management for more details',
+      );
     }
     if (volume && (volume < 0 || volume > 1)) {
       throw new Error('Volume must be between 0 and 1, or undefined');